Display suffix in tags
Instead of adding the suffix at the very end, we should display the
suffix exactly where we cut the string.

This means that instead of having the following result:

```html
<p>Lorem ipsum dolor</p>…
```

We should have:

```html
<p>Lorem ipsum dolor…</p>
```

In that case the ellipsis does not end up in its own orphan line when
rendered.